Added Specialty Subs

+++ SandWhich/annotations/src/main/java/com/pluralsight/annotation/menu/option/MenuOption.java
@@ -9,7 +9,7 @@ import java.lang.annotation.Target;
 @Target(ElementType.FIELD)
 public @interface MenuOption {
     String key();
-    int order();
+    double order();
     String delimiter() default " - ";
     String overrideName() default "";
 }
